+++ packages/core/src/constants/gemini.ts
@@ -1,0 +1,15 @@
+// API Endpoints
+export const ENDPOINT_GEMINI_API = 'https://generativelanguage.googleapis.com/v1';
+
+// gemini model
+export const MODEL_GEMINI_2_0_FLASH = 'gemini-2.0-flash';
+export const MODEL_GEMINI_2_0_FLASH_LITE = 'gemini-2.0-flash-lite';
+export const MODEL_GEMINI_1_5_FLASH = 'gemini-1.5-flash';
+
+
+// Vision support for Gemini models
+export const GEMINI_VISION_SUPPORTED_MODELS = [
+  MODEL_GEMINI_2_0_FLASH,
+  MODEL_GEMINI_2_0_FLASH_LITE,
+  MODEL_GEMINI_1_5_FLASH,
+];
