+++ packages/core/src/constants/voiceEngines.ts
@@ -1,0 +1,4 @@
+export const VOICE_VOX_API_URL = 'http://localhost:50021';
+export const VOICEPEAK_API_URL = 'http://localhost:20202';
+export const AIVIS_SPEECH_API_URL = 'http://localhost:10101';
+export const NIJI_VOICE_API_URL = 'https://api.nijivoice.com/api/platform/v1';
