+++ packages/core/src/constants/youtube.ts
@@ -1,0 +1,1 @@
+export const ENDPOINT_YOUTUBE_API = 'https://youtube.googleapis.com/youtube/v3';
